requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# PyTorch and ML
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84

# Storage (optional S3 support)
boto3==1.29.7
boto3-stubs[s3]==1.29.7

# Development tools
black==23.11.0
isort==5.12.0
ruff==0.1.6
mypy==1.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
